+++ packages/sonata-api/src/index.ts
@@ -1,6 +1,7 @@
 export * from '@sonata-api/access-control'
 export * from '@sonata-api/api'
 export * from '@sonata-api/common'
+export * from '@sonata-api/entrypoint'
 export * from '@sonata-api/http'
 export * from '@sonata-api/mailing'
 export * from '@sonata-api/node-http'