+++ packages/security/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @sonata-api/security
 
+## 1.0.5
+
+### Patch Changes
+
+- Include new packages
+- Updated dependencies
+  - @sonata-api/access-control@1.0.6
+  - @sonata-api/common@0.1.8
+  - @sonata-api/api@1.0.7
+
+## 1.0.4
+
+### Patch Changes
+
+- Add new packages
+- Updated dependencies
+  - @sonata-api/access-control@1.0.5
+  - @sonata-api/common@0.1.7
+  - @sonata-api/api@1.0.6
+
 ## 1.0.3
 
 ### Patch Changes